Add a typed idl2json (#13)
# Motivation
It would be nice to have type names in records and variants. If a did
file is available, this can be done on a best-effort basis.

# Changes
- Bump the candid version to 0.8.1
- Move the untyped idl2json to another file in the library
- Add a typed idl2json to the library
- Add an example to illustrate how the typed conversion works, printing
intermediate values.
- Add a `rust_toolchain.toml` as the code doesn't compile with whatever
Rust I was using at first.
- Added clippy lints to ensure that the code cannot panic with
`panic()`, `unwrap()` or `expect()`. The conversion should never fail.
- Added a function to convert internal types to IDLTypes, because
`#[derive(Candid)]` provides the internal type.

# Tests
- Added a single "happy path" test for decoding binary candid to JSON
with a given schema.
- The same test is repeated with derived type information.