fix: Support vector (#24)
# Motivation
The idl2json command line assumes the common case of a single candid
structure:
```
$ echo "( record {} )" | idl2json 
{}
```
however there may be multiple or no structures. This is currently
unsupported. Effectively this means that the output becomes an array
with zero or more entries. This can be done without breaking backwards
compatibility by emitting zero or more JSON structures:
```
$ echo "()" | idl2json 
$ echo "( record {} )" | idl2json 
{}
$ echo "( record {}, record {} )" | ./target/debug/idl2json 
{}
{}
```
# Changes
- Support an arbitrary number of `IDLValue`s
- Refactor to take a string as input and output of the main library
function, to make it testable.
- Add tests
- Bump the cargo version
- Address clippy issues in the existing code

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>