Parse service (#26)
# Motivation
We wish to be able to convert proposal arguments to JSON. This requires:
- Canister arguments are of type IDLArgs which is a vector of types,
rather than one IDLType. Assuming that the argument is just one IDLType
leads to errors that have to be handled downstream. It is better to fix
it here, at source.
- Canister arguments often refer to types defined in the did file. The
did file is an IDLProg which has a declarations section, where these
types are defined. Thus we need optional support for providing one (or
more) IDLProgs and looking in them for type definitions. If a type is
not found, we fall back to unannotated conversion to JSON.
- We need to be able to extract init args from a did file. There is
already code for this in the nns-dapp repo that we can steal.
- We also need tests, to make sure that we don't break backwards
compatibility more than necessary as we add these features.

# Changes
- All of the above, basically. Lots of tests and the functionality
described there.
- Added support for the above functionality to the idl2json CLI.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>