Cli bytes as (#33)
# Motivation
Proposals contain didc wrapped in didc, so they need to be parsed, then
the payload bytes need to be parsed again. `didc` takes hex as input, so
the payload blob is needed as hex, not an array of ints. The `idl2json`
library already supports this, but the option is not exposed in the CLI.

# Changes
* Expose the idl2json bytes formatting option in the CLI.
Note: The "long bytes" option is not exposed, as the clap for that is
complicated and this is not a priority.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>