Set the idl2json version in the workspace (#39)
# Motivation
The `idl2json_cli` needs to use the same version of candid as
`idl2json`. Currently, this is enforced only by this comment:
```
candid = "0.8.1" # Should match the version in the lib.
```
We can provide a stronger guarantee that they are the same by providing
the candid version in the workspace.

# Changes
- Define the candid version in the workspace Cargo.toml.
- Use the workspace candid version in `idl2json` and `idl2json_cli`, but
not `yaml2candid`. The last currently uses a different version. Perhaps
it _should_ use the same but it does not.

# Tests
Existing tests should suffice.

Co-authored-by: Max Murphy <max@winning.black>