Use the workspace candid for yaml2candid (#40)
# Motivation
`yaml2candid` does not currently use the workspace candid version. It
would be better if it did, so that all the `idl2json` tools used the
same version of `candid`.

# Changes
- Use the workspace version of `candid` for `yaml2candid`.

# Tests
CI should suffice for most purposes.

For the `yaml2candid` CLI, the change does not take effect until the
release version is bumped. I have tested updating the release version
locally.

Co-authored-by: Max Murphy <max@winning.black>