Yaml2json types (#95)
# Motivation
`yaml2candid` is very incomplete and still in some places buggy.
Specifically, it does not support many Candid types and the opt test
vectors (and implementation that was guided by those test vectors) are
wrong.

# Changes
- Restructure the main `yaml2json` type match so that the compiler will
complain when not all types are matched.
- Add YAML to Candid conversions for almost all missing Candid types.
Service and class are still unsupported. I am running out of time and
these are both complicated and not needed for my immediate goal.
- Add lots of tests.
- Fix the existing test vector that didn't include `opt` in the candid.

# Tests
- Fairly comprehensive unit tests are included for each individual type.
- Test coverage was checked with Tarpaulin; the tarpaulin report is
generated by CI and can be checked.
- Most but not all error branches are tested.

# TODO in future
- Ensure that failures have useful error messages.
- Use proptests to ensure that no corner cases have been overlooked.
- Perhaps get more test vectors from real-life use to ensure that the
test vectors are consistent with conventional usage.